Move medal components to a folder, refactor to function components
