+++ src/profile/components/Profile/Medals/medals.less
@@ -1,0 +1,34 @@
+.crownContainer {
+  position: relative;
+  min-height: 80px;
+}
+
+.crown {
+  padding: 3px 8px;
+  width: 100px;
+  vertical-align: bottom;
+  position: absolute;
+  bottom: 0;
+  box-sizing: border-box;
+}
+
+.medalContainer {
+  display: grid;
+  grid-template-columns: auto;
+  margin: 0 30px;
+  transition: 1s ease-in 0.2s;
+  text-align: center;
+  scroll-snap-align: start;
+  width: 100%;
+  box-sizing: border-box;
+}
+
+.medal {
+  width: 100px;
+  box-sizing: border-box;
+}
+
+.yearRange {
+  font-weight: bold;
+  text-align: center;
+}
