+++ src/profile/less/profile.less
@@ -33,58 +33,11 @@
   margin-bottom: 20px;
 }
 
-.committeeMedal {
-  width: 100px;
-
-  @media screen and (max-width: @owMobileBreakpoint) {
-    width: 36px;
-  }
-}
-
-.committeeCrownContainer {
-  position: relative;
-  min-height: 80px;
-}
-
-.committeeCrown {
-  padding: 3px 8px;
-  width: 85px;
-  vertical-align: bottom;
-  position: absolute;
-  bottom: 0;
-
-  @media screen and (max-width: @owMobileBreakpoint) {
-    width: 30px;
-    padding: 1px 3px;
-  }
-}
-
-.committeeRange {
-  font-weight: bold;
-  text-align: center;
-}
-
-.medalContainer {
-  display: grid;
-  grid-template-columns: auto;
-  margin: 0 30px;
-}
-
-.sideScroll {
-  overflow-x: scroll;
-}
-
 .groupWrapper {
   display: grid;
   grid-template-columns: repeat(2, 275px);
   justify-content: center;
   grid-gap: 20px;
-}
-
-.medalGrid {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  justify-items: center;
 }
 
 .infoType,
